permute: move program-cache regression tests into test_permute.py (reviewer nit)
